Bikes edit - complete

+++ AperoApp.API/Models/Bike.cs
@@ -12,7 +12,7 @@ namespace AperoApp.API.Models
         public int Gears { get; set; }
         public string Color { get; set; }
         public int WheelSize { get; set; }
-        public int Weight { get; set; }
+        public float Weight { get; set; }
         public string Type { get; set; }
         public bool Basket { get; set; }
         public string Description { get; set; }
